requirements.txt
argcomplete==3.3.0
boto3==1.34.114
botocore==1.34.114
certifi==2024.2.2
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
durationpy==0.6
hjson==3.1.0
holidays==0.49
idna==3.7
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
numpy==1.26.4
pandas==2.2.2
placebo==0.9.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
s3transfer==0.10.1
six==1.16.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.4
troposphere==4.8.0
tzdata==2024.1
urllib3==1.26.6
Werkzeug==3.0.3
zappa==0.59.0
